chemical_project/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
pandas==3.0.0
psycopg2-binary==2.9.10
reportlab==4.2.6
numpy==2.4.1
